[41] - date calculation service setup
