+++ src/Auctionify.API/ClientApp/src/app/components/buyer/dashboard/dashboard.component.ts
@@ -5,6 +5,7 @@ import { MatSnackBar } from '@angular/material/snack-bar';
 
 import { Client, LotModel } from 'src/app/web-api-client';
 import { RemoveFromWatchlistComponent } from '../../general/remove-from-watchlist/remove-from-watchlist.component';
+import { DateCalculationService } from 'src/app/services/date-calculation/date-calculation.service';
 
 @Component({
     selector: 'app-dashboard',
@@ -16,7 +17,8 @@ export class DashboardComponent implements OnInit {
     constructor(
         private apiClient: Client,
         private snackBar: MatSnackBar,
-        private dialog: Dialog
+        private dialog: Dialog,
+        private dateCalculationService: DateCalculationService
     ) {}
 
     ngOnInit(): void {
@@ -33,29 +35,14 @@ export class DashboardComponent implements OnInit {
         startDate: Date | null,
         endDate: Date | null
     ): number | null {
-        if (!startDate || !endDate) {
-            return null;
-        }
-
-        const start = new Date(startDate);
-        const end = new Date(endDate);
-        const timeDifference = end.getTime() - start.getTime();
-        return Math.ceil(timeDifference / (1000 * 3600 * 24));
+        return this.dateCalculationService.calculateDaysLeftActive(
+            startDate,
+            endDate
+        );
     }
 
     calculateDaysLeftUpcoming(startDate: Date | null): number | null {
-        if (!startDate) {
-            return null;
-        }
-
-        const now = new Date();
-        const start = new Date(startDate);
-        if (start <= now) {
-            return 0;
-        }
-
-        const timeDifference = start.getTime() - now.getTime();
-        return Math.ceil(timeDifference / (1000 * 3600 * 24));
+        return this.dateCalculationService.calculateDaysLeftUpcoming(startDate);
     }
 
     handleLotWatchlist(lot: LotModel) {
